İş ilanı CRUD eklendi
İş ilanı CRUD eklendi

+++ src/main/java/com/kodlamaio/hrms/dataAccess/abstracts/BusinessPositionDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.kodlamaio.hrms.entites.concretes.BusinessPosition;
 
 
-public interface BusinessPositionDao extends JpaRepository<BusinessPosition, Integer>{
+public interface BusinessPositionDao extends JpaRepository<BusinessPosition, Short>{
 
 }
